Remove connector
The connector never work for me, and a lot of people on Gitter are complaining on this too for at least a week now.

I suggest we remove the server for now, until its more stable / has more room. It would save a lot of trouble shooting on moneyd for new users.